KafkaSampleProducer/requirements.txt
kafka-python ~= 2.0.2
quoters ~= 0.30